% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_round.R
\name{adapt_round}
\alias{adapt_round}
\title{a helper function for rounding numbers, big and small.}
\usage{
adapt_round(x)
}
\arguments{
\item{x}{numeric vector of numbers.}
}
\value{
character vector of rounded numbers
}
\description{
a helper function for rounding numbers, big and small.
}
\examples{

adapt_round(c(1.256, 11.22, 100.230))

}
