% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{conduct an analysis of binary or survival outcomes}
\usage{
analyze(data_orig, data_mult = NULL, formula, control,
  nested_models = TRUE, analysis_type = NULL,
  include.unadjusted = TRUE, variables_to_test = NULL,
  person_years = 1000)
}
\arguments{
\item{data_orig}{the unimputed data. This is all you need if your data don't have missing values.}

\item{data_mult}{a list of imputed data. This is needed if the original data have missing values.}

\item{formula}{a formula object with outcomes on the left hand side and main exposures on the right hand side.}

\item{control}{a list of formulas with each item in the list representing a new control model and containing a function with control variables on the right hand side (see examples).}

\item{nested_models}{should the models specified in control be nested?}

\item{analysis_type}{one of 'rel_risk' or 'prop_haz'.}

\item{include.unadjusted}{should an unadjusted model be included?}

\item{variables_to_test}{a character vector of variables that should be tested in addition to the main exposure.}

\item{person_years}{only relevant if the outcome is time-to-event. The denominator for incidence rates. Default is incidence rate per 1000 person-years.}
}
\description{
conduct an analysis of binary or survival outcomes
}
