% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_models.R
\name{bind_models}
\alias{bind_models}
\title{turn a list of tidy models into a data frame of tidy models.}
\usage{
bind_models(models, model_formulas, descriptives = NULL)
}
\arguments{
\item{models}{the models}

\item{model_formulas}{formulas for the models}

\item{descriptives}{a tibble of descriptive statistics}
}
\description{
turn a list of tidy models into a data frame of tidy models.
}
