% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_descriptives.R
\name{cmp_descr}
\alias{cmp_descr}
\alias{cmp_descr.list}
\alias{cmp_descr.rel_risk_outcome}
\alias{cmp_descr.prop_haz_outcome}
\title{compute descriptive statistics for a solo table}
\usage{
cmp_descr(outcome, exposure, data, N)

\method{cmp_descr}{list}(outcome, exposure, data, N)

\method{cmp_descr}{rel_risk_outcome}(outcome, exposure, data, N)

\method{cmp_descr}{prop_haz_outcome}(outcome, exposure, data, N)
}
\arguments{
\item{outcome}{dependent variable}

\item{exposure}{exposure variable}

\item{data}{data frame}

\item{N}{denominator for incidence rate calculation}
}
\description{
compute descriptive statistics for a solo table
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: compute descriptive statistics over a list

\item \code{rel_risk_outcome}: compute descriptive statistics for a relative risk analysis

\item \code{prop_haz_outcome}: compute descriptive statistics for a binary outcome
}}

