% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{cmp_estms.list}
\alias{cmp_estms.list}
\title{modeling function for hazard ratios}
\usage{
\method{cmp_estms}{list}(fit, exposure, data)
}
\arguments{
\item{fit}{a fitted model}

\item{exposure}{the exposure variable}

\item{data}{a data frame for model fitting.}
}
\value{
model estimates
}
\description{
modeling function for hazard ratios
}
