% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codify.R
\name{codify}
\alias{codify}
\title{organize the output of a solo object into a table.}
\usage{
codify(solo_input, type = c("long", "wide"), filler_value = " ")
}
\arguments{
\item{solo_input}{an object returned by the analyze function}

\item{type}{one of "long", "wide" or c("long","wide").}

\item{filler_value}{the value that goes into blank cells}
}
\description{
organize the output of a solo object into a table.
}
