% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_binary_descriptives.R
\name{compute_binary_descriptives}
\alias{compute_binary_descriptives}
\title{compute size and prevalence of a binary outcome for exposure groups}
\usage{
compute_binary_descriptives(data, exposure, outcome)
}
\arguments{
\item{data}{a data set}

\item{exposure}{the exposure}

\item{outcome}{the outcome}
}
\description{
compute size and prevalence of a binary outcome for exposure groups
}
