% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_survival_descriptives.R
\name{compute_survival_descriptives}
\alias{compute_survival_descriptives}
\title{compute descriptive statistics for survival outcomes}
\usage{
compute_survival_descriptives(data, exposure, status, time, N = 1000)
}
\arguments{
\item{data}{a data frame}

\item{exposure}{a character vector}

\item{status}{a character vector}

\item{time}{a character}

\item{N}{the denominator for rates, e.g., rate per N=1000 person-years}
}
\description{
compute descriptive statistics for survival outcomes
}
