% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model.list}
\alias{fit_model.list}
\title{fit a model with the given formula}
\usage{
\method{fit_model}{list}(formula, exposure, data,
  variables_to_test = NULL, keep.model = TRUE)
}
\arguments{
\item{formula}{a modeling formula}

\item{exposure}{the exposure variable}

\item{data}{a data frame for model fitting.}

\item{variables_to_test}{character vector of variables to test.}

\item{keep.model}{should the model object be kept?}
}
\value{
a fitted model
}
\description{
fit a model with the given formula
}
