% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{modeling function for hazard ratios}
\usage{
fit_model(formula, data, tidy)
}
\arguments{
\item{formula}{a modeling formula}

\item{data}{a data frame for model fitting.}

\item{tidy}{logical. Should the model be tidied before it is returned?}
}
\value{
a fitted proportional hazards model
}
\description{
modeling function for hazard ratios
}
