% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_work_data_frame.R
\name{gee_work.data.frame}
\alias{gee_work.data.frame}
\title{Generate a nice table with prevalence ratios}
\usage{
\method{gee_work}{data.frame}(data, model_formulas, exposure)
}
\arguments{
\item{data}{a data frame or mids object}

\item{model_formulas}{a list of model formulas}

\item{exposure}{a character indicating the column name of the exposure.}
}
\value{
a list of gee models
}
\description{
Generate a nice table with prevalence ratios
}
