% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_work.R
\name{gee_work}
\alias{gee_work}
\title{Generate a nice table with prevalence ratios}
\usage{
gee_work(model_formulas, data, exposure)
}
\arguments{
\item{model_formulas}{a list of model formulas}

\item{data}{a data frame}

\item{exposure}{a character indicating the column name of the exposure.}
}
\value{
a list of gee models
}
\description{
Generate a nice table with prevalence ratios
}
