% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_table.R
\name{hr_table}
\alias{hr_table}
\title{Generate a publication-ready table with hazard ratios}
\usage{
hr_table(data, exposure, time, status, control,
  include.unadjusted = TRUE, include.descriptive = TRUE,
  person_years_denominator = 1000, return_data = FALSE)
}
\arguments{
\item{data}{a data frame}

\item{exposure}{a named character vector. Names are variable labels. Values are column names in data.}

\item{time}{a named character vector. Names are labels for time variables. Values are column names of the time columns in data.}

\item{status}{a named character vector. Names are labels for status variables. Values are column names of the status columns in data.}

\item{control}{a list of character vectors. The ith item in the list should contain column names of variables that will be added as control variables for the ith model. Naming the character vectors will result in the names being used as labels for control variables in the footnote of the table.}

\item{include.unadjusted}{logical. Should unadjusted prevalence ratios be presented in the table?}

\item{include.descriptive}{logical. Should event rates be presented in the table?}

\item{person_years_denominator}{integer. Incidence rates will be computed per this number.}

\item{return_data}{logical. Should the table data be returned instead of the table?}
}
\description{
Generate a publication-ready table with hazard ratios
}
