% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_elements.R
\name{list_elements}
\alias{list_elements}
\title{helper function to list things.}
\usage{
list_elements(x, end_string = "and ", sep_string = "=",
  col_string = ", ", use_names = FALSE, lower = TRUE)
}
\arguments{
\item{x}{a vector of elements}

\item{end_string}{the character that precedes the last item in x}

\item{sep_string}{the string that links names to values}

\item{col_string}{the string that precedes items other than the first and last in x}

\item{use_names}{if names are used, the names of x are listed along with values in x}

\item{lower}{if true, then all characters are printed in lower case.}
}
\value{
a string that lists the elements of x
}
\description{
helper function to list things.
}
