% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tidy_models_data_frame.R
\name{make_tidy_models.data.frame}
\alias{make_tidy_models.data.frame}
\title{Generate a nice table with ratios}
\usage{
\method{make_tidy_models}{data.frame}(data, model_formulas, exposure)
}
\arguments{
\item{data}{a data frame or mids object}

\item{model_formulas}{a list of model formulas}

\item{exposure}{a character indicating the column name of the exposure.}
}
\value{
a list of tidy model summaries
}
\description{
Generate a nice table with ratios
}
