% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_control_variables.R
\name{parse_control_variables}
\alias{parse_control_variables}
\title{parse control variables}
\usage{
parse_control_variables(control_variables, data)
}
\arguments{
\item{control_variables}{the control variables}

\item{data}{a data frame containing the control variables.}
}
\description{
parse control variables
}
