% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_formula.R
\name{parse_formula}
\alias{parse_formula}
\title{parse a formula and set its class to be solo compatible}
\usage{
parse_formula(formula, outcome)
}
\arguments{
\item{formula}{a model formula}

\item{outcome}{the outcome variable}
}
\description{
parse a formula and set its class to be solo compatible
}
