% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_table.R
\name{pr_table}
\alias{pr_table}
\title{Generate a nice table with prevalence ratios}
\usage{
pr_table(data, exposure, outcome, control, include.unadjusted = TRUE,
  include.descriptive = TRUE, return_data = FALSE,
  collapse_footer = FALSE)
}
\arguments{
\item{data}{a data frame}

\item{exposure}{a named character vector. Names are variable labels. Values are column names in the data.}

\item{outcome}{a named character vector. Names are outcome labels. Values are column names of the outcomes in the data.}

\item{control}{a list of character vectors. The ith item in the list should contain column names of variables that will be added as control variables for the ith model. Naming the character vectors will result in the names being used as labels for control variables in the footnote of the table.}

\item{include.unadjusted}{logical. Should unadjusted prevalence ratios be presented in the table?}

\item{include.descriptive}{logical. Should the prevalence of the outcome be presented in the table?}

\item{return_data}{logical. Should the table data be returned instead of the table?}

\item{collapse_footer}{logical. Should the model footnote be collapsed?}
}
\description{
Generate a nice table with prevalence ratios
}
\examples{
library(titanic)
library(magrittr)
library(dplyr)
library(gt)

data <- titanic_train \%>\%
  dplyr::select(Survived, Pclass, Sex, Age, SibSp, Fare) \%>\%
  as_tibble() \%>\%
  dplyr::mutate(
    Pclass=factor(
      Pclass,
      levels=c(1,2,3),
      labels=c("1st Class","2nd Class","3rd Class")
    )
  ) \%>\%
  na.omit()

tst_output <- pr_table(
  data=data,
  exposure=c("Passenger class"='Pclass'),
  outcome=c("Survival"='Survived'),
  control=list(
    c("Passenger gender"="Sex"),
    c("Passenger age"="Age",
      "Ticket price"="Fare"),
    c("Gender by age interaction"="Sex:Age")
  ),
  include.unadjusted=TRUE,
  include.descriptive=TRUE
)
}
