% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_solo.R
\name{print.solo_object}
\alias{print.solo_object}
\title{print a solo object}
\usage{
\method{print}{solo_object}(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{arguments passed to other functions}
}
\description{
print a solo object
}
