% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrub_data.R
\name{scrub_data}
\alias{scrub_data}
\title{check inputs and return the original (og) data.}
\usage{
scrub_data(data, all_vars, exposure, control)
}
\arguments{
\item{data}{the data}

\item{all_vars}{all of the variables specified for modeling}

\item{exposure}{the designated exposure variable for model fitting}

\item{control}{list of model parameters}
}
\description{
check inputs and return the original (og) data.
}
