% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrub_variables.R
\name{scrub_variables}
\alias{scrub_variables}
\title{identify unique variables in a character vector of model terms}
\usage{
scrub_variables(...)
}
\arguments{
\item{...}{character strings or vectors of model terms}
}
\description{
identify unique variables in a character vector of model terms
}
