% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_exposure.R
\name{set_exposure_class}
\alias{set_exposure_class}
\title{grab the exposure variables from a formula}
\usage{
set_exposure_class(exposure, data)
}
\arguments{
\item{exposure}{the exposure variable}

\item{data}{a dataset with labelled columns}
}
\description{
grab the exposure variables from a formula
}
