% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_outcome.R
\name{set_outcome_class}
\alias{set_outcome_class}
\title{grab the outcome variables from a formula}
\usage{
set_outcome_class(outcome, new_class, data)
}
\arguments{
\item{outcome}{outcome variable}

\item{new_class}{new class for the outcome}

\item{data}{a dataset with labelled columns}
}
\description{
grab the outcome variables from a formula
}
