% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_hypotheses}
\alias{test_hypotheses}
\title{test statistical hypotheses from a solo model}
\usage{
test_hypotheses(model, exposure, data, variables_to_test = NULL)
}
\arguments{
\item{model}{a fitted model object}

\item{exposure}{exposure variable}

\item{data}{data frame}

\item{variables_to_test}{character vector of variables to test}
}
\description{
test statistical hypotheses from a solo model
}
