% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_footer.R
\name{write_footer}
\alias{write_footer}
\title{writes a footnote describing the models in a prevalence table}
\usage{
write_footer(model_predictors, control)
}
\arguments{
\item{model_predictors}{a list describing model predictors.}

\item{control}{a list describing control parameters.}
}
\value{
a character value of a table footnote
}
\description{
writes a footnote describing the models in a prevalence table
}
