% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_predictors.R
\name{gather_predictors}
\alias{gather_predictors}
\title{helper function to translate control objects into lists of model formulas}
\usage{
gather_predictors(control)
}
\arguments{
\item{control}{a list of character vectors.}
}
\value{
a set of parameters for regression models
}
\description{
helper function to translate control objects into lists of model formulas
}
